crew-algorithms/pyproject.toml
[tool.poetry]
name = "crew-algorithms"
version = "0.1.0"
description = ""
authors = ["Lingyu Zhang <lingyu.zhang@duke.edu>, Zhengran Ji <zhengran.ji@duke.edu>, Boyuan Chen <boyuan.chen@duke.edu>"]
readme = "README.md"
packages = [{include = "crew_algorithms"}]
repository = "https://github.com/generalroboticslab/CREW/crew-algorithms"

[tool.poetry.dependencies]
python = ">=3.10.7,<=3.10.11"
numpy = "^1.23.5"
antialiased-cnns = "^0.3"
torch = "2.2.0"
torchrl = "0.3.0"
torchvision = "^0.17"
pytorch-ignite = "^0.4.12"
hydra-core = "^1.3.2"
attrs = "^22.2.0"
protobuf = "<3.20"
cattrs = "^22.2.0"
wandb = "^0.15.3"
mlagents-envs = {git = "https://github.com/lingyu98/ml-agents.git", subdirectory = "ml-agents-envs"}
sortedcontainers = "^2.4.0"
openai = {extras = ["wandb"], version = "^0.27.8"}
openai-whisper  = {git = "https://github.com/openai/whisper.git"}
pyaudio = "0.2.11"
matplotlib = ">=3.8.2"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pre-commit = "^3.2.2"

[tool.poetry.group.docs.dependencies]
furo = "^2023.3.27"
sphinx-remove-toctrees = "^0.0.3"
docutils = "^0.19"
sphinxcontrib-video = "^0.2.0"

[tool.isort]
profile = 'black'
known_third_party = ["wandb"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
